Initial commit
This is an extension of the the Sunshine v2 project from Udacity.
What has been added since "hello_world" branch:
RecyclerView with CardView for displaying 10-day forecast.
Test data for only one day in the 10-day forecast.
Ability to display day of week, type of weather, degrees in Fahrenheit,
and degrees in Celcius